Update build process in .github/workflows/build.yml
Updated the build and upload process in the GitHub Actions workflow file to target the `release-ranger` package instead of `dc`. The changes affect both the Linux and MacOS builds. The move to this package ensures proper versioning and information display during runtime. Also, it streams physiques renaming of output files to 'rer' for consistency.